Clean up OID handling changes
This is mostly just refactoring Nirgals patch to use a standard
capability function. I also short-circuited default_with_oid check since it
will always show false in >=PG12, so this saves a round trip. This also adds
a comment in the config file that show oids is no longer supported. This
could lead to a slight difference in running instalations, but since it
does not affect any code/behavior, not going to bump the config version.